dom/tests/mochitest/geolocation/test_allowCurrent.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=478911
-->
<head>
  <title>Test for getCurrentPosition </title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="geolocation_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=478911">Mozilla Bug 478911</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

var hasAccepted = false;

function successCallback(position) {
  ok(hasAccepted, "Ensure that accept was pressed");
  check_geolocation(position);
  SimpleTest.finish();
}

function accept() {
  hasAccepted = true;
  clickNotificationButton(kAcceptButton);
}

SimpleTest.waitForExplicitFinish();

var options = {
    maximumAge: 0,
};

// one-shot position requests
navigator.geolocation.getCurrentPosition(successCallback, null, options);

setTimeout(accept, 50);

</script>
</pre>
</body>
</html>
